feat: authorization header check

+++ src/modules/user/user.controller.ts
@@ -2,7 +2,8 @@ import { v4 as uuidv4 } from 'uuid';
 import type { FastifyReply, FastifyRequest } from 'fastify';
 
 import { prisma } from '@/libs/prisma';
-import {
+import type { TAuthorizationHeaderObject } from '@/modules/_common/schemas/auth';
+import type {
   // userInputSchema,
   // userSchema,
   UserInputType,
@@ -17,9 +18,16 @@ import {
 import { PrismaClientKnownRequestError } from '@prisma/client/runtime';
 
 export const createUserHandler = async (
-  request: FastifyRequest<{ Body: UserInputType }>,
+  request: FastifyRequest<{ Body: UserInputType; Headers: TAuthorizationHeaderObject }>,
   reply: FastifyReply,
 ): Promise<void> => {
+  const { authorization } = request.headers;
+  if (authorization === undefined) {
+    reply.unauthorized();
+    return;
+  }
+  request.log.info(`authorization header: ${authorization}`);
+
   const id = uuidv4();
   const date = new Date();
 
@@ -86,9 +94,16 @@ export const getUsersHandler = async (
 };
 
 export const deleteUserHandler = async (
-  request: FastifyRequest<{ Params: GetUserParamsType }>,
+  request: FastifyRequest<{ Params: GetUserParamsType; Headers: TAuthorizationHeaderObject }>,
   reply: FastifyReply,
 ): Promise<void> => {
+  const { authorization } = request.headers;
+  if (authorization === undefined) {
+    reply.unauthorized();
+    return;
+  }
+  request.log.info(`authorization header: ${authorization}`);
+
   try {
     const user = await prisma.user.delete({
       where: {
@@ -109,9 +124,20 @@ export const deleteUserHandler = async (
 };
 
 export const patchUserHandler = async (
-  request: FastifyRequest<{ Params: PatchUserParamsType; Body: PatchUserRequestBodyType }>,
+  request: FastifyRequest<{
+    Params: PatchUserParamsType;
+    Body: PatchUserRequestBodyType;
+    Headers: TAuthorizationHeaderObject;
+  }>,
   reply: FastifyReply,
 ): Promise<void> => {
+  const { authorization } = request.headers;
+  if (authorization === undefined) {
+    reply.unauthorized();
+    return;
+  }
+  request.log.info(`authorization header: ${authorization}`);
+
   const data: { name?: string; email?: string; rank?: number; updatedAt: Date } = {
     updatedAt: new Date(),
   };
@@ -140,9 +166,20 @@ export const patchUserHandler = async (
 };
 
 export const putUserHandler = async (
-  request: FastifyRequest<{ Params: PutUserParamsType; Body: PutUserRequestBodyType }>,
+  request: FastifyRequest<{
+    Params: PutUserParamsType;
+    Body: PutUserRequestBodyType;
+    Headers: TAuthorizationHeaderObject;
+  }>,
   reply: FastifyReply,
 ): Promise<void> => {
+  const { authorization } = request.headers;
+  if (authorization === undefined) {
+    reply.unauthorized();
+    return;
+  }
+  request.log.info(`authorization header: ${authorization}`);
+
   const { id } = request.params;
   const { name, email, birthdate, rank } = request.body;
   const date = new Date();
